feat: optimize battle logic
